Test Platform - UI: Update Preview and filter options in question list
