+++ test-platform-ui/src/app/administration/questions/(components)/preview/data.ts
@@ -1,0 +1,10 @@
+/** Sample filter options: */
+
+export const _filterOpts = [
+  { key: 'gilad', value: 'Gilad Gray' },
+  { key: 'jason', value: 'Jason Killian' },
+  { key: 'antoine', value: 'Antoine Llorca' },
+  { key: 'gilad1', value: 'Gilad Gray' },
+  { key: 'jason1', value: 'Jason Killian' },
+  { key: 'antoine1', value: 'Antoine Llorca' },
+];
